refactored App adding screens
