+++ components/common/ScreenInfo.tsx
@@ -2,8 +2,9 @@ import { View, Text, Dimensions, StyleSheet } from 'react-native';
 import { getStatusBarHeight } from 'react-native-status-bar-height';
 import { useSelector } from 'react-redux';
 
-import { State } from '../redux';
-import { colors } from '../config';
+import { State } from '../../redux';
+import { colors } from '../../config';
+import { screen } from '../../helpers';
 
 export default function ScreenInfo() {
   const theme = useSelector((state: State) => state.theme.theme);
@@ -19,7 +20,7 @@ export default function ScreenInfo() {
   return (
     <View>
       <Text style={styles.para}>{`screen is ${JSON.stringify(
-        Dimensions.get('screen'),
+        screen,
         null,
         2,
       )}`}</Text>
