+++ components/screens/Home.tsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native';
+import ScreenInfo from '../common/ScreenInfo';
+import Counter from '../common/Counter';
+
+const Home = () => (
+  <View>
+    <ScreenInfo />
+    <Counter />
+  </View>
+);
+
+export default Home;
